// packages/frontend/src/types/schemas/prisma-generated/inputTypeSchemas/NestedEnumEmailStatusWithAggregatesFilterSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { EmailStatusSchema } from './EmailStatusSchema';
import { NestedIntFilterSchema } from './NestedIntFilterSchema';
import { NestedEnumEmailStatusFilterSchema } from './NestedEnumEmailStatusFilterSchema';

export const NestedEnumEmailStatusWithAggregatesFilterSchema: z.ZodType<Prisma.NestedEnumEmailStatusWithAggregatesFilter> = z.object({
  equals: z.lazy(() => EmailStatusSchema).optional(),
  in: z.lazy(() => EmailStatusSchema).array().optional(),
  notIn: z.lazy(() => EmailStatusSchema).array().optional(),
  not: z.union([ z.lazy(() => EmailStatusSchema),z.lazy(() => NestedEnumEmailStatusWithAggregatesFilterSchema) ]).optional(),
  _count: z.lazy(() => NestedIntFilterSchema).optional(),
  _min: z.lazy(() => NestedEnumEmailStatusFilterSchema).optional(),
  _max: z.lazy(() => NestedEnumEmailStatusFilterSchema).optional()
}).strict();

export default NestedEnumEmailStatusWithAggregatesFilterSchema;